Restructure Project
