+++ ObsidianSQL.library/IConnection.cs
@@ -1,0 +1,10 @@
+namespace ObsidianSQL.library
+{
+    public interface IConnection
+    {
+        public IDatabase[] Databases { get; }
+        public void Connect();
+        public void Disconnect();
+        public int ExecuteQuery(string query);
+    }
+}
