+++ ObsidianSQL.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ObsidianSQL.server", "Obsid
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObsidianSQL.library", "ObsidianSQL.library\ObsidianSQL.library.csproj", "{50AD4BBC-11AB-4B8A-B7D4-4A3D1038656E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObisidanSQL.impl.mysql", "ObisidanSQL.impl.mysql\ObisidanSQL.impl.mysql.csproj", "{404A17CA-91CF-4774-8364-054EA2A28B86}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObisidanSQL.library.mysql", "ObisidanSQL.library.mysql\ObisidanSQL.library.mysql.csproj", "{404A17CA-91CF-4774-8364-054EA2A28B86}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
