+++ ObsidianSQL/ObsidianSQL.server.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ObisidanSQL.impl.mysql\ObisidanSQL.impl.mysql.csproj" />
+    <ProjectReference Include="..\ObisidanSQL.library.mysql\ObisidanSQL.library.mysql.csproj" />
     <ProjectReference Include="..\ObsidianSQL.library\ObsidianSQL.library.csproj" />
   </ItemGroup>
