+++ ObsidianSQL/src/ObsidianSQL.cs
@@ -3,10 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using ObsidianSQL.library.controller;
-using ObisidanSQL.impl.mysql.controller;
+using ObsidianSQL.server.controller;
 
-namespace ObsidianSQL.server.src
+namespace ObsidianSQL.server
 {
     class ObsidianSQL : IDisposable
     {
