+++ ObsidianSQL/src/Router.cs
@@ -4,9 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ObsidianSQL.library;
-using ObsidianSQL.server.src.exceptions;
+using ObsidianSQL.server.exceptions;
+using ObsidianSQL.server.http;
 
-namespace ObsidianSQL.server.src
+namespace ObsidianSQL.server
 {
     class Router
     {
