+++ ObsidianSQL/src/controller/IController.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using ObsidianSQL.server.http;
 
-namespace ObsidianSQL.library.controller
+namespace ObsidianSQL.server.controller
 {
     public interface IController
     {
