+++ ObsidianSQL/src/exceptions/RouteNotFoundException.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ObsidianSQL.server.src.exceptions
+namespace ObsidianSQL.server.exceptions
 {
     class RouteNotFoundException : Exception
     {
