+++ ObsidianSQL/src/http/IRequest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ObsidianSQL.library
+namespace ObsidianSQL.server.http
 {
     public interface IRequest
     {
